+++ src/__fixtures__/object-patterns/invalid-comments.ts
