+++ src/__fixtures__/object-patterns/valid-comments.ts
